requirements.txt
langchain>=0.0.207
openai>=0.27.0
chromadb>=0.3.21
langgraph>=0.1.0
streamlit>=1.25.0
python-dotenv
python-docx>=0.8.11
PyPDF2>=3.0.0